content/base/test/test_copypaste.html

<!DOCTYPE HTML>
<html>
<!--
-->
<head>
  <title>Test for copy/paste</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=524975">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

function testCopyPaste () {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  var webnav = window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
                     .getInterface(Components.interfaces.nsIWebNavigation)

  var docShell = webnav.QueryInterface(Components.interfaces.nsIDocShell);

  var documentViewer = docShell.contentViewer
                               .QueryInterface(Components.interfaces.nsIContentViewerEdit);

  var clipboard = Components.classes["@mozilla.org/widget/clipboard;1"]
                            .getService(Components.interfaces.nsIClipboard);

  var textarea = document.getElementById('input');

  function copySelectionToClipboard() {
    documentViewer.copySelection();
    is(clipboard.hasDataMatchingFlavors(["text/unicode"], 1,1), true);
    is(clipboard.hasDataMatchingFlavors(["text/html"], 1,1), true);
  }
  function copyToClipboard(node) {
    textarea.blur();
    clipboard.emptyClipboard(1);
    var sel = window.getSelection();
    sel.removeAllRanges();
    var r = document.createRange();
    r.selectNode(node);
    window.getSelection().addRange(r);
    copySelectionToClipboard();
  }
  function copyRangeToClipboard(startNode,startIndex,endNode,endIndex) {
    textarea.blur();
    clipboard.emptyClipboard(1);
    var sel = window.getSelection();
    sel.removeAllRanges();
    var r = document.createRange();
    r.setStart(startNode,startIndex)
    r.setEnd(endNode,endIndex)
    window.getSelection().addRange(r);
    copySelectionToClipboard();
  }
  function copyChildrenToClipboard(id) {
    textarea.blur();
    clipboard.emptyClipboard(1);
    window.getSelection().selectAllChildren(document.getElementById(id));
    copySelectionToClipboard();
  }
  function getClipboardData(mime) {
    var transferable = Components.classes['@mozilla.org/widget/transferable;1']
                                 .createInstance(Components.interfaces.nsITransferable);
    transferable.addDataFlavor(mime);
    clipboard.getData(transferable, 1);
    var data = {};
    transferable.getTransferData(mime, data, {}) ;
    return data;
  }
  function testClipboardValue(mime, expected) {
    var data = getClipboardData(mime);
    is (data.value == null ? data.value :
        data.value.QueryInterface(Components.interfaces.nsISupportsString).data,
      expected,
      mime + " value in the clipboard");
    return data.value;
  }
  function testPasteText(expected) {
    textarea.value="";
    textarea.focus();
    textarea.QueryInterface(Components.interfaces.nsIDOMNSEditableElement)
            .editor.paste(1);
    is(textarea.value, expected, "value of the textarea after the paste");
  }
  function testSelectionToString(expected) {
    is(window.getSelection().toString().replace(/\r\n/g,"\n"), expected, "Selection.toString");
  }
  function testInnerHTML(id, expected) {
    var value = document.getElementById(id).innerHTML;
    is(value, expected, id + ".innerHTML");
  }
  function testEmptyChildren(id) {
    copyChildrenToClipboard(id);
    testSelectionToString("");
    testClipboardValue("text/unicode", null);
    testClipboardValue("text/html", null);
    testPasteText("");
  }

  copyChildrenToClipboard("draggable");
  testSelectionToString("This is a draggable bit of text.");
  testClipboardValue("text/unicode",
                     "This is a draggable bit of text.");
  testClipboardValue("text/html",
                     "<div id=\"draggable\" title=\"title to have a long HTML line\">This is a <em>draggable</em> bit of text.</div>");
  testPasteText("This is a draggable bit of text.");

  copyChildrenToClipboard("alist");
  testSelectionToString(" bla\n\n    foo\n    bar\n\n");
  testClipboardValue("text/unicode", "bla\n\n    foo\n    bar\n\n");
  testClipboardValue("text/html", "<div id=\"alist\">\n    bla\n    <ul>\n      <li>foo</li>\n      \n      <li>bar</li>\n    </ul>\n  </div>");
  testPasteText("bla\n\n    foo\n    bar\n\n");

  copyChildrenToClipboard("blist");
  testSelectionToString(" mozilla\n\n    foo\n    bar\n\n");
  testClipboardValue("text/unicode", "mozilla\n\n    foo\n    bar\n\n");
  testClipboardValue("text/html", "<div id=\"blist\">\n    mozilla\n    <ol>\n      <li>foo</li>\n      \n      <li>bar</li>\n    </ol>\n  </div>");
  testPasteText("mozilla\n\n    foo\n    bar\n\n");

  copyChildrenToClipboard("clist");
  testSelectionToString(" mzla\n\n    foo\n        bazzinga!\n    bar\n\n");
  testClipboardValue("text/unicode", "mzla\n\n    foo\n        bazzinga!\n    bar\n\n");
  testClipboardValue("text/html", "<div id=\"clist\">\n    mzla\n    <ul>\n      <li>foo<ul>\n        <li>bazzinga!</li>\n      </ul></li>\n      \n      <li>bar</li>\n    </ul>\n  </div>");
  testPasteText("mzla\n\n    foo\n        bazzinga!\n    bar\n\n");

  copyChildrenToClipboard("div4");
  testSelectionToString(" Tt t t ");
  testClipboardValue("text/unicode", "Tt t t ");
  testClipboardValue("text/html", "<div id=\"div4\">\n  T<textarea>t t t</textarea>\n</div>");
  testInnerHTML("div4", "\n  T<textarea>t t t</textarea>\n");
  testPasteText("Tt t t ");

  copyChildrenToClipboard("div5");
  testSelectionToString(" T ");
  testClipboardValue("text/unicode", "T ");
  testClipboardValue("text/html", "<div id=\"div5\">\n  T<textarea>     </textarea>\n</div>");
  testInnerHTML("div5", "\n  T<textarea>     </textarea>\n");
  testPasteText("T ");

  copyRangeToClipboard($("div6").childNodes[0],0, $("div6").childNodes[1],1);
  testSelectionToString("");
// START Disabled due to bug 564688
if (false) {
  testClipboardValue("text/unicode", "");
  testClipboardValue("text/html", "");
}
// END Disabled due to bug 564688
  testInnerHTML("div6", "div6");

  copyRangeToClipboard($("div7").childNodes[0],0, $("div7").childNodes[0],4);
  testSelectionToString("");
// START Disabled due to bug 564688
if (false) {
  testClipboardValue("text/unicode", "");
  testClipboardValue("text/html", "");
}
// END Disabled due to bug 564688
  testInnerHTML("div7", "div7");

  copyRangeToClipboard($("div8").childNodes[0],0, $("div8").childNodes[0],4);
  testSelectionToString("");
// START Disabled due to bug 564688
if (false) {
  testClipboardValue("text/unicode", "");
  testClipboardValue("text/html", "");
}
// END Disabled due to bug 564688
  testInnerHTML("div8", "div8");

  copyRangeToClipboard($("div9").childNodes[0],0, $("div9").childNodes[0],4);
  testSelectionToString("div9");
  testClipboardValue("text/unicode", "div9");
  testClipboardValue("text/html", "div9");
  testInnerHTML("div9", "div9");

  copyToClipboard($("div10"));
  testSelectionToString("");
  testInnerHTML("div10", "div10");

  copyToClipboard($("div10").firstChild);
  testSelectionToString("");

  copyRangeToClipboard($("div10").childNodes[0],0, $("div10").childNodes[0],1);
  testSelectionToString("");

  copyRangeToClipboard($("div10").childNodes[1],0, $("div10").childNodes[1],1);
  testSelectionToString("");

  // ============ copy/paste test from/to a textarea

  var val = "1\n 2\n  3";
  textarea.value=val;
  textarea.select();
  textarea.editor.copy();
  
  textarea.value="";
  textarea.editor.paste(1);
  is(textarea.value, val);
  textarea.value="";

  SimpleTest.finish();
}


SimpleTest.waitForExplicitFinish();

addLoadEvent(testCopyPaste);

</script>
</pre>
<div>

  <div id="draggable" title="title to have a long HTML line">This is a <em>draggable</em> bit of text.</div>
  <textarea id="input" cols="40" rows="10"></textarea>

  <div id="alist">
    bla
    <ul>
      <li>foo</li>
      <li style="display: none;">baz</li>
      <li>bar</li>
    </ul>
  </div>

  <div id="blist">
    mozilla
    <ol>
      <li>foo</li>
      <li style="display: none;">baz</li>
      <li>bar</li>
    </ol>
  </div>

  <div id="clist">
    mzla
    <ul>
      <li>foo<ul>
        <li>bazzinga!</li>
      </ul></li>
      <li style="display: none;">baz</li>
      <li>bar</li>
    </ul>
  </div>

<div id="div4">
  T<textarea>t t t</textarea>
</div>

<div id="div5">
  T<textarea>     </textarea>
</div>

<div id="div6" style="display:none"></div>
<script>
var x = $("div6")
x.appendChild(document.createTextNode('di'))
x.appendChild(document.createTextNode('v6'))
</script>

<div id="div7" style="display:none">div7</div>
<div id="div8" style="visibility:hidden">div8</div>
<div style="visibility:hidden"><div id="div9" style="visibility:visible">div9</div></div>
<div style="visibility:hidden"><div><div><div id="div10"></div></div></div></div>
<script>
var x = $("div10")
x.appendChild(document.createTextNode('div'))
x.appendChild(document.createTextNode('10'))
</script>


</div>
</body>
</html>
